refactor(server): extract the output tool into src/server/handlers/output-tools.ts
get_output and handleGetOutput move off the class verbatim (`this.` ->
`ctx.`). handleGetOutput is not a test seam, so no delegate stays behind.
